+++ hyperdbg/hprdbgkd/header/debugger/core/Debugger.h
@@ -228,7 +228,7 @@ PDEBUGGER_EVENT
 DebuggerGetEventByTag(UINT64 Tag);
 
 BOOLEAN
-DebuggerRemoveEvent(UINT64 Tag, BOOLEAN InputFromVmxRoot);
+DebuggerRemoveEvent(UINT64 Tag, BOOLEAN PoolManagerAllocatedMemory);
 
 BOOLEAN
 DebuggerQueryDebuggerStatus();
@@ -250,22 +250,20 @@ DebuggerParseAction(PDEBUGGER_GENERAL_ACTION          Action,
 
 BOOLEAN
 DebuggerParseEventsModification(PDEBUGGER_MODIFY_EVENTS DebuggerEventModificationRequest,
-                                BOOLEAN                 InputFromVmxRoot);
+                                BOOLEAN                 InputFromVmxRoot,
+                                BOOLEAN                 PoolManagerAllocatedMemory);
 
 BOOLEAN
 DebuggerTerminateEvent(UINT64 Tag, BOOLEAN InputFromVmxRoot);
 
-BOOLEAN
-DebuggerEnableOrDisableAllEvents(BOOLEAN IsEnable);
-
-BOOLEAN
-DebuggerRemoveAllEvents(BOOLEAN InputFromVmxRoot);
-
-BOOLEAN
-DebuggerTerminateAllEvents(BOOLEAN InputFromVmxRoot);
-
 UINT32
 DebuggerEventListCount(PLIST_ENTRY TargetEventList);
+
+VOID
+DebuggerClearAllEvents(BOOLEAN InputFromVmxRoot, BOOLEAN PoolManagerAllocatedMemory);
+
+BOOLEAN
+DebuggerEnableOrDisableAllEvents(BOOLEAN IsEnable);
 
 UINT32
 DebuggerEventListCountByCore(PLIST_ENTRY TargetEventList, UINT32 TargetCore);
@@ -287,6 +285,9 @@ DebuggerQueryStateEvent(UINT64 Tag);
 
 BOOLEAN
 DebuggerDisableEvent(UINT64 Tag);
+
+BOOLEAN
+DebuggerClearEvent(UINT64 Tag, BOOLEAN InputFromVmxRoot, BOOLEAN PoolManagerAllocatedMemory);
 
 VOID
 DebuggerPerformActions(PROCESSOR_DEBUGGING_STATE * DbgState, DEBUGGER_EVENT * Event, DEBUGGER_TRIGGERED_EVENT_DETAILS * EventTriggerDetail);