+++ hyperdbg/script-eval/code/Functions.c
@@ -693,51 +693,75 @@ ScriptEngineFunctionInterlockedCompareExchange(
 }
 
 /**
- * @brief Implementation of enable_event function
+ * @brief Implementation of event_enable function
  *
- * @param Tag
- * @param ImmediateMessagePassing
- * @param Value
+ * @param EventId
+ *
  * @return VOID
  */
 VOID
-ScriptEngineFunctionEnableEvent(UINT64  Tag,
-                                BOOLEAN ImmediateMessagePassing,
-                                UINT64  Value)
+ScriptEngineFunctionEventEnable(UINT64 EventId)
 {
 #ifdef SCRIPT_ENGINE_USER_MODE
     ShowMessages("err, enabling events is not possible in user-mode\n");
 #endif // SCRIPT_ENGINE_USER_MODE
 
 #ifdef SCRIPT_ENGINE_KERNEL_MODE
-    if (!DebuggerEnableEvent(Value + DebuggerEventTagStartSeed))
+    if (!DebuggerEnableEvent(EventId + DebuggerEventTagStartSeed))
     {
-        LogInfo("Invalid tag id (%x)", Value);
+        LogInfo("Invalid tag id (%x)", EventId);
     }
 #endif // SCRIPT_ENGINE_KERNEL_MODE
 }
 
 /**
- * @brief Implementation of disable_event function
+ * @brief Implementation of event_disable function
  *
- * @param Tag
- * @param ImmediateMessagePassing
- * @param Value
+ * @param EventId
+ *
  * @return VOID
  */
 VOID
-ScriptEngineFunctionDisableEvent(UINT64  Tag,
-                                 BOOLEAN ImmediateMessagePassing,
-                                 UINT64  Value)
+ScriptEngineFunctionEventDisable(UINT64 EventId)
 {
 #ifdef SCRIPT_ENGINE_USER_MODE
     ShowMessages("err, disabling events is not possible in user-mode\n");
 #endif // SCRIPT_ENGINE_USER_MODE
 
 #ifdef SCRIPT_ENGINE_KERNEL_MODE
-    if (!DebuggerDisableEvent(Value + DebuggerEventTagStartSeed))
+    if (!DebuggerDisableEvent(EventId + DebuggerEventTagStartSeed))
     {
-        LogInfo("Invalid tag id (%x)", Value);
+        LogInfo("Invalid tag id (%x)", EventId);
+    }
+#endif // SCRIPT_ENGINE_KERNEL_MODE
+}
+
+/**
+ * @brief Implementation of event_clear function
+ *
+ * @param EventId
+ *
+ * @return VOID
+ */
+VOID
+ScriptEngineFunctionEventClear(UINT64 EventId)
+{
+#ifdef SCRIPT_ENGINE_USER_MODE
+    ShowMessages("err, disabling events is not possible in user-mode\n");
+#endif // SCRIPT_ENGINE_USER_MODE
+
+#ifdef SCRIPT_ENGINE_KERNEL_MODE
+
+    BOOLEAN PoolManagerAllocatedMemory = FALSE;
+
+    if (g_KernelDebuggerState && EnableInstantEventMechanism)
+    {
+        PoolManagerAllocatedMemory = TRUE;
+    }
+
+    if (!DebuggerClearEvent(EventId + DebuggerEventTagStartSeed, VmFuncVmxGetCurrentExecutionMode(), PoolManagerAllocatedMemory))
+    {
+        LogInfo("Invalid tag id (%x)", EventId);
     }
 #endif // SCRIPT_ENGINE_KERNEL_MODE
 }